Move KH2 BOBs in Kh2Field
